TASK-47872 Enhance Getting Started Widget Performances Display Prior to this change, the widget was initialized using Javascript which can lead to bad FrontEnd performances by multiplying the REST calls and JS execution to initialize Vue application. This change will make the application HTML built in Server side since it's a simple widget display
