+++ webapp/portlet/src/main/webapp/WEB-INF/portlet.xml
@@ -728,33 +728,13 @@
       <title>Top Bar Notification</title>
     </portlet-info>
   </portlet>
-  
+
   <portlet>
     <portlet-name>GettingStarted</portlet-name>
     <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
     <init-param>
       <name>portlet-view-dispatched-file-path</name>
-      <value>/html/gettingStarted.html</value>
-    </init-param>
-    <init-param>
-      <name>use-js-manager</name>
-      <value>true</value>
-    </init-param>
-    <init-param>
-      <name>js-manager-jsModule</name>
-      <value>PORTLET/social-portlet/GettingStarted</value>
-    </init-param>
-    <init-param>
-      <name>prefetch.resources</name>
-      <value>true</value>
-    </init-param>
-    <init-param>
-      <name>prefetch.resource.bundles</name>
-      <value>locale.portlet.social.GettingStartedPortlet</value>
-    </init-param>
-    <init-param>
-      <name>prefetch.resource.rest</name>
-      <value>/portal/rest/getting-started</value>
+      <value>/WEB-INF/jsp/gettingStarted.jsp</value>
     </init-param>
     <supports>
       <mime-type>text/html</mime-type>
