--- webapp/portlet/src/main/webapp/html/gettingStarted.html
@@ -1,5 +1,0 @@
-<div class="VuetifyApp">
-  <div id="GettingStartedPortlet">
-    <v-cacheable-dom-app cache-id="GettingStartedPortlet"></v-cacheable-dom-app>
-  </div>
-</div>
