--- webapp/portlet/src/main/webapp/vue-apps/getting-started/initComponents.js
@@ -1,9 +1,0 @@
-import ExoGettingStarted from './components/ExoGettingStarted.vue';
-
-const components = {
-  'exo-getting-started': ExoGettingStarted,
-};
-
-for (const key in components) {
-  Vue.component(key, components[key]);
-}
